refactor interrupt handling
